Upgrade to use std::future::Future.
This upgrades to use the stabalised `std::future::Future`, and the
`futures-0.3` crate for the convenient `TryFuture` for futures returning
a result. This allows easy use with async-await. It means the minimum
Rust version would become 1.39. The change in types is a breaking semver
change.

This depends on `futures-core` rather than the `futures` crate because
we're only using the `TryFuture` crate and it brings the number of
dependencies down significantly.

This adds an extra dependency on `pin-project` to handle the pinning of
the inner future in `ResponseFuture`. It'd be possible to avoid this
dependency, but I didn't trust myself to get the pinning rules right.
It's a fairly lightweight dependency.