Upgrade to use std::future::Future. (#15)
* Upgrade to use std::future::Future.

This upgrades to use the stabalised `std::future::Future`, and the
`futures-0.3` crate for the convenient `TryFuture` for futures returning
a result. This allows easy use with async-await. It means the minimum
Rust version would become 1.39. The change in types is a breaking semver
change.

This depends on `futures-core` rather than the `futures` crate because
we're only using the `TryFuture` crate and it brings the number of
dependencies down significantly.

This adds an extra dependency on `pin-project` to handle the pinning of
the inner future in `ResponseFuture`. It'd be possible to avoid this
dependency, but I didn't trust myself to get the pinning rules right.
It's a fairly lightweight dependency.

* Upgrade to tokio-0.2 in the tests.

* Upgrade rand.

This adds an extra dependency on `rand_xorshift` for the tests, as the
PRNG has been moved out of the main crate.

* Upgrade parking_lot.

* Update futures bench.

* cargo update

* Add futures-support feature.

The `futures` module can be optional for users who do not needing it,
saving them a few dependencies. This feature is default on, for
compatibility with the current API.

I would have liked to call this `futures`, but can't because features
and dependencies share a namespace.

* Update CHANGELOG/README.

* Update CircleCI for minimum Rust version.